Add(tests): Extract `CountingQuoteRepository` test double for shared usage across modules
- Moved reusable test double implementation into a dedicated `test_double.rs` module under the `quote` domain.
- Updated tests in `backfill/quote_history` to use `CountingQuoteRepository` from the new module.
- Excluded `test_double.rs` from code coverage reporting via `codecov.yml` to avoid distortion from unimplemented methods.
